pyproject.toml
[tool.poetry]
name = "abilian-core"
version = "0.11.17"
description = "A framework for enterprise applications (CRM, ERP, collaboration...), based on Flask and SQLAlchemy"
authors = ["Abilian SAS"]
license = "LGPL"
readme = "README.rst"
repository = "https://github.com/abilian/abilian-core"
packages = [{ include = "abilian", from = "src" }]
include = ["src/**/*"]


[tool.poetry.dependencies]
python = ">=3.6.1,<4"

# Flask and friends
Flask = "^1.0"
Flask-Assets = ">=0.12"
Flask-Mail = ">=0.9.1"
Flask-Migrate = ">=2.0"
Flask-Login = ">=0.4"
Flask-Talisman = ">=0.6"

# Pin werkzeug and flask-babel for now
werkzeug = "<1"
Flask-Babel = ">=0.11,<2"

# flask-sqlachemy 2.2 breaks our apps.
flask-sqlalchemy = "<=2.1"

# Currently broken w/ 0.13
Flask-WTF = ">=0.12,<0.13"

# Currently broken w/ 2.0
webassets = "< 2"

# Forms add-ons
WTForms-Alchemy = ">=0.12"
wtforms-sqlalchemy = "*"

# pathlib is included in python 3.4, and very useful
pathlib = "*"

# can parse isoformat / used in a jinja filter
python-dateutil = "^2.4.0"

# Databases / persistence
SQLAlchemy = "^1.1"
alembic = ">=0.9"

# requirement of jinja2; > 0.21 has an important fix for string format. See
# https://github.com/mitsuhiko/markupsafe/issues/28
MarkupSafe = ">=0.21"

# Indexing
Whoosh = "^2.5"

# Task queue
celery = "~4"
redis = "~3"

# Security
bleach = ">=2"
bcrypt = "*"

# low-level tools
deprecated = "*"
hyperlink = "*"

# Used by DeferredJS (TODO: remove)
lxml = "*"

# Used by the logging config loader (TODO: remove)
PyYAML = "*"

# Used by the connection audit
pygeoip = "*"

# Progres bar, used by some scripts
tqdm = "*"

# Used by the Sentry extension
sentry-sdk = { version="*", extras=["flask"] }

# for debug toolbar
sqlparse = "*"

# antivirus interface
clamd = "*"

# Content detection / transformation
python-magic = "*"
pillow = "*"

# Reporting
pandas = ">=0.17"

#
# Assets management (might not be necessary if distributing compiled assets)
#
jsmin = "*"
cssmin = "*"
closure = "*"
devtools = "*"

[tool.poetry.dev-dependencies]
Flask-DebugToolbar = ">=0.10"

# For running tests (pytest and friends)
pytest = ">=2.4.0"
pytest-flask = "^1"
pytest-xdist = "*"
pytest-randomly = "*"
pytest-repeat = "*"
pytest-cov = "*"
typeguard = "*"
nox = "*"
tox = "*"

# Static code analysis: flake* and friends
flake8 = "*"
flake8-assertive = "*"
flake8-blind-except = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-isort = "*"
flake8-logging-format = "*"
flake8-mutable = "*"
flake8-pytest = "*"
flake8-super-call = "*"
flake8-tidy-imports = "*"
mastool = "*"
mccabe = "*"

mypy = "*"

# Later
# flake8-pep3101
# this one tries too hard
# pep8-naming

# Development tools
pre-commit = "*"

# Formatting
isort = "*"
docformatter = "*"
black = { version = "*", allow-prereleases = true }

# To build docs
# (Doesn't work w/ sphinx 1.4.5)
Sphinx = ">=1.5.5"

# Standard theme for readthedocs.com
sphinx-rtd-theme = "*"

# Used to generate changelogs
gitchangelog = "^3.0"

# Used by functional tests
requests = "*"

# Used by some tests
html5lib = "*"

# Check README
restructuredtext_lint = "*"

# Temp (workaround poetry issue)
docutils = "*"

# Doc (to try later)
# portray = "^1.3"

# Check deps
safety = "^1.9"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
